requirements.txt
streamlit
opencv-python
numpy
pillow
scikit-learn
sqlalchemy
requests
torch
torchvision
